Added base health

+++ TypeAndSlash/Assets/Code/Interface/IDamageable.cs
@@ -1,7 +1,8 @@
-﻿namespace Code.Interface
+﻿
+namespace Code.Interface
 {
     public interface IDamageable
     {
-        public void TakeDamage(int damage);
+        public void TakeDamage(float damage);
     }
 }
